agents/pyproject.toml
[project]
name = "enterprise-agent-backend"
version = "0.1.0"
description = "엔터프라이즈 멀티 에이전트 지휘자(Orchestrator) 봇 서버"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "mcp",
    "fastapi",
    "uvicorn",
    "langchain",
    # 추가로 필요한 OpenAI, Anthropic 등의 패키지를 여기에 명시합니다.
    "langchain-openai>=1.1.12",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "langgraph>=1.1.3",
    "tavily-python>=0.5.0",
    "common_util",
]

[tool.uv.sources]
common_util = { path = "../common_util", editable = true }
